refactor(system-settings): 重构系统设置页面为模块化目录结构
将系统设置从扁平的文件结构（SystemSettings.tsx + SystemModalContent.tsx）重构为独立的目录模块 src/renderer/pages/settings/system/，提取 DirInputItem 和 PreferenceRow 可复用组件，降低耦合度，提升可维护性
