+++ src/renderer/pages/settings/system/components/DirInputItem.tsx
@@ -1,0 +1,61 @@
+import { Button, Form, Tooltip } from '@arco-design/web-react';
+import { FolderOpen } from '@icon-park/react';
+import React from 'react';
+import { useTranslation } from 'react-i18next';
+import { ipcBridge } from '@/common';
+
+/**
+ * 目录选择输入组件 / Directory selection input component
+ * 用于选择和显示系统目录路径 / Used for selecting and displaying system directory paths
+ */
+export default function DirInputItem({ label, field }: IDirInputItemProps) {
+  const { t } = useTranslation();
+
+  return (
+    <Form.Item label={label} field={field}>
+      {(_value, form) => {
+        const currentValue = form.getFieldValue(field) || '';
+
+        const handlePick = () => {
+          ipcBridge.dialog.showOpen
+            .invoke({
+              defaultPath: currentValue,
+              properties: ['openDirectory', 'createDirectory'],
+            })
+            .then((res) => {
+              if (res?.success && res.data && !res.data.canceled && res.data.filePaths.length > 0) {
+                form.setFieldValue(field, res.data.filePaths[0]);
+              }
+            })
+            .catch((error) => {
+              console.error('Failed to open directory dialog:', error);
+            });
+        };
+
+        return (
+          <div className='aion-dir-input h-[32px] flex items-center rounded-8px border border-solid border-transparent pl-14px bg-[var(--fill-0)]'>
+            <Tooltip content={currentValue || t('settings.dirNotConfigured')} position='top'>
+              <div className='flex-1 min-w-0 text-13px text-foreground truncate '>{currentValue || t('settings.dirNotConfigured')}</div>
+            </Tooltip>
+            <Button
+              type='text'
+              style={{ borderLeft: '1px solid var(--color-border-2)', borderRadius: '0 8px 8px 0' }}
+              icon={<FolderOpen theme='outline' size='18' fill={'var(--foreground)'} />}
+              onClick={(e) => {
+                e.stopPropagation();
+                handlePick();
+              }}
+            />
+          </div>
+        );
+      }}
+    </Form.Item>
+  );
+}
+
+interface IDirInputItemProps {
+  /** 标签文本 / Label text */
+  label: string;
+  /** 表单字段名 / Form field name */
+  field: string;
+}
